+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add ability to define colors for widgets' attributes with RGB values or Hex color codes.
 * Extend pager widget to mark hidden desktops that contain windows.
 * Fix calculation of cpu usage.
+* Add widget for commands. Shows the output of configured cli commands.
 
 # v0.3.0